вынес тесты меню в отдельный файл и класс
